Split attribute.rb
Now that the public APIs of the entities involved in an attribute are
stable, it is preferrable to locate each entity definition in its own
file.

Also, the visibility of those entities changes from private to public,
first because Ruby demands it with this structure of files, but also
because the entities' APIs are stable.
