+++ lib/sheetah/attribute_types/composite.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+require_relative "value"
+
+module Sheetah
+  module AttributeTypes
+    class Composite
+      def initialize(composite:, scalars:)
+        @composite_type = composite
+        @values = scalars.map { |scalar| Value.new(scalar) }
+      end
+
+      def compile(container)
+        container.composite(@composite_type, @values.map(&:type))
+      end
+
+      def each_column
+        return enum_for(:each_column) { @values.size } unless block_given?
+
+        @values.each_with_index do |value, index|
+          yield index, value.required
+        end
+
+        self
+      end
+
+      def freeze
+        @values.freeze
+        @values.each(&:freeze)
+        super
+      end
+    end
+  end
+end
